model demo and refactoring
