+++ kinodata/model/shared/__init__.py
@@ -1,2 +1,2 @@
-from .gine import GINE
+from .gine import GINE, LigandGINE
 from .set_transformer import SetAttentionBlock, MultiheadAttentionBlock
