+++ scripts/train_sparse_transformer.py
@@ -9,15 +9,10 @@ import wandb
 import kinodata.configuration as configuration
 from kinodata.training import train
 from kinodata.data.data_module import make_kinodata_module
-from kinodata.model.complex_transformer import ComplexTransformer
+from kinodata.model.complex_transformer import ComplexTransformer, make_model
 from kinodata.types import NodeType
 from kinodata.data.dataset import apply_transform_instance_permament
 from kinodata.transform.to_complex_graph import TransformToComplexGraph
-
-
-def make_model(config: configuration.Config):
-    cls = partial(ComplexTransformer, config)
-    return config.init(cls)
 
 
 if __name__ == "__main__":
